requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
numpy==2.1.3
pandas